Migrated replication_spec migration code to be encompassed by bi connector migration
