--- mongodbatlas/resource_mongodbatlas_advanced_cluster_v0.go
@@ -1,242 +1,0 @@
-package mongodbatlas
-
-import (
-	"bytes"
-	"context"
-	"log"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
-)
-
-func resourceMongoDBAtlasAdvancedClusterV0() *schema.Resource {
-	log.Println("[INFO] Found MongoDB Cluser state v0; migrating to v1")
-	return &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"project_id": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
-			},
-			"cluster_id": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"backup_enabled": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				Computed: true,
-			},
-			"bi_connector": {
-				Type:     schema.TypeList,
-				Optional: true,
-				Computed: true,
-				MaxItems: 1,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"enabled": {
-							Type:     schema.TypeBool,
-							Optional: true,
-							Computed: true,
-						},
-						"read_preference": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Computed: true,
-						},
-					},
-				},
-			},
-			"cluster_type": {
-				Type:     schema.TypeString,
-				Required: true,
-			},
-			"connection_strings": clusterConnectionStringsSchema(),
-			"create_date": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"disk_size_gb": {
-				Type:     schema.TypeFloat,
-				Optional: true,
-				Computed: true,
-			},
-			"encryption_at_rest_provider": {
-				Type:     schema.TypeString,
-				Optional: true,
-				Computed: true,
-			},
-			"labels": {
-				Type:     schema.TypeSet,
-				Optional: true,
-				Set: func(v interface{}) int {
-					var buf bytes.Buffer
-					m := v.(map[string]interface{})
-					buf.WriteString(m["key"].(string))
-					buf.WriteString(m["value"].(string))
-					return HashCodeString(buf.String())
-				},
-				Computed: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"key": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Computed: true,
-						},
-						"value": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Computed: true,
-						},
-					},
-				},
-			},
-			"mongo_db_major_version": {
-				Type:      schema.TypeString,
-				Optional:  true,
-				Computed:  true,
-				StateFunc: formatMongoDBMajorVersion,
-			},
-			"mongo_db_version": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"name": {
-				Type:     schema.TypeString,
-				Required: true,
-				ForceNew: true,
-			},
-			"paused": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				Default:  false,
-			},
-			"pit_enabled": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				Computed: true,
-			},
-			"replication_specs": {
-				Type:     schema.TypeSet,
-				Required: true,
-				Elem: &schema.Resource{
-					Schema: map[string]*schema.Schema{
-						"id": {
-							Type:     schema.TypeString,
-							Computed: true,
-						},
-						"num_shards": {
-							Type:         schema.TypeInt,
-							Optional:     true,
-							Default:      1,
-							ValidateFunc: validation.IntBetween(1, 50),
-						},
-						"region_configs": {
-							Type:     schema.TypeSet,
-							Required: true,
-							Elem: &schema.Resource{
-								Schema: map[string]*schema.Schema{
-									"analytics_specs": advancedClusterRegionConfigsSpecsSchema(),
-									"auto_scaling": {
-										Type:     schema.TypeList,
-										MaxItems: 1,
-										Optional: true,
-										Computed: true,
-										Elem: &schema.Resource{
-											Schema: map[string]*schema.Schema{
-												"disk_gb_enabled": {
-													Type:     schema.TypeBool,
-													Optional: true,
-													Computed: true,
-												},
-												"compute_enabled": {
-													Type:     schema.TypeBool,
-													Optional: true,
-													Computed: true,
-												},
-												"compute_scale_down_enabled": {
-													Type:     schema.TypeBool,
-													Optional: true,
-													Computed: true,
-												},
-												"compute_min_instance_size": {
-													Type:     schema.TypeString,
-													Optional: true,
-													Computed: true,
-												},
-												"compute_max_instance_size": {
-													Type:     schema.TypeString,
-													Optional: true,
-													Computed: true,
-												},
-											},
-										},
-									},
-									"backing_provider_name": {
-										Type:     schema.TypeString,
-										Optional: true,
-									},
-									"electable_specs": advancedClusterRegionConfigsSpecsSchema(),
-									"priority": {
-										Type:     schema.TypeInt,
-										Required: true,
-									},
-									"provider_name": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-									"read_only_specs": advancedClusterRegionConfigsSpecsSchema(),
-									"region_name": {
-										Type:     schema.TypeString,
-										Required: true,
-									},
-								},
-							},
-						},
-						"container_id": {
-							Type: schema.TypeMap,
-							Elem: &schema.Schema{
-								Type: schema.TypeString,
-							},
-							Computed: true,
-						},
-						"zone_name": {
-							Type:     schema.TypeString,
-							Optional: true,
-							Default:  "ZoneName managed by Terraform",
-						},
-					},
-				},
-				Set: replicationSpecsHashSet,
-			},
-			"root_cert_type": {
-				Type:     schema.TypeString,
-				Optional: true,
-				Computed: true,
-			},
-			"state_name": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"termination_protection_enabled": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				Computed: true,
-			},
-			"version_release_system": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				Default:      "LTS",
-				ValidateFunc: validation.StringInSlice([]string{"LTS", "CONTINUOUS"}, false),
-			},
-			"advanced_configuration": clusterAdvancedConfigurationSchema(),
-		},
-	}
-}
-
-func resourceMongoDBAtlasAdvancedClusterStateUpgradeV0(ctx context.Context, rawState map[string]interface{}, meta interface{}) (map[string]interface{}, error) {
-	log.Println("[INFO] Found MongoDB Cluser state v0; migrating to v1")
-
-	return rawState, nil
-}
